requirements.txt
fastapi
uvicorn
pytesseract
Pillow
python-multipart
python-dateutil